// server/routes/modules/callManagerModule.types.ts
export type CallManagerModuleDeps = {
  addCallsToThreadInMicroBatches: any;
  analyzeCallTranscript: any;
  analyzeTranscriptQualification: any;
  callDispatcher: any;
  calculateNextAvailableCallTime: any;
  checkAdminAccess: any;
  checkFlyVoiceProxyHealth: any;
  checkIfStoreOpen: any;
  columnIndexToLetter: any;
  googleSheets: any;
  isAdmin: any;
  isAuthenticatedCustom: any;
  parseHoursToStructured: any;
  reconcileOrphanedCallSessions: any;
  storage: any;
  syncAgentSettingsFromElevenLabs: any;
  voiceProxyServer: any;
};